requirements.txt
fastapi==0.75.2
uvicorn==0.17.6
sqlalchemy==1.4.36
psycopg2==2.9.3
python-decouple==3.6
alembic==1.7.7
fastapi-utils==0.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
email-validator==1.2.1
fastapi-mail==1.0.8
python-slugify==6.1.2